+++ Collector/ProfileClient.php
@@ -3,9 +3,11 @@
 namespace Http\HttplugBundle\Collector;
 
 use Http\Client\Common\FlexibleHttpClient;
+use Http\Client\Exception\HttpException;
 use Http\Client\HttpAsyncClient;
 use Http\Client\HttpClient;
 use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
 
 /**
  * The ProfileClient decorates any client that implement both HttpClient and HttpAsyncClient interfaces to gather target
@@ -28,11 +30,17 @@ class ProfileClient implements HttpClient, HttpAsyncClient
     private $collector;
 
     /**
+     * @var Formatter
+     */
+    private $formatter;
+
+    /**
      * @param HttpClient|HttpAsyncClient $client    The client to profile. Client must implement both HttpClient and
      *                                              HttpAsyncClient interfaces.
      * @param Collector                  $collector
+     * @param Formatter                  $formatter
      */
-    public function __construct($client, Collector $collector)
+    public function __construct($client, Collector $collector, Formatter $formatter)
     {
         if (!($client instanceof HttpClient && $client instanceof HttpAsyncClient)) {
             throw new \RuntimeException(sprintf(
@@ -45,6 +53,7 @@ class ProfileClient implements HttpClient, HttpAsyncClient
         }
         $this->client = $client;
         $this->collector = $collector;
+        $this->formatter = $formatter;
     }
 
     /**
@@ -52,9 +61,18 @@ class ProfileClient implements HttpClient, HttpAsyncClient
      */
     public function sendAsyncRequest(RequestInterface $request)
     {
-        $this->collectRequestInformations($request);
+        $stack = $this->collector->getCurrentStack();
+        $this->collectRequestInformations($request, $stack);
 
-        return $this->client->sendAsyncRequest($request);
+        return $this->client->sendAsyncRequest($request)->then(function (ResponseInterface $response) use ($stack) {
+            $this->collectResponseInformations($response, $stack);
+
+            return $response;
+        }, function (\Exception $exception) use ($stack) {
+            $this->collectExceptionInformations($exception, $stack);
+
+            throw $exception;
+        });
     }
 
     /**
@@ -62,22 +80,67 @@ class ProfileClient implements HttpClient, HttpAsyncClient
      */
     public function sendRequest(RequestInterface $request)
     {
-        $this->collectRequestInformations($request);
+        $stack = $this->collector->getCurrentStack();
+        $this->collectRequestInformations($request, $stack);
 
-        return $this->client->sendRequest($request);
+        try {
+            $response = $this->client->sendRequest($request);
+
+            $this->collectResponseInformations($response, $stack);
+
+            return $response;
+        } catch (\Exception $e) {
+            $this->collectExceptionInformations($e, $stack);
+
+            throw $e;
+        }
     }
 
     /**
      * @param RequestInterface $request
+     * @param Stack|null       $stack
      */
-    private function collectRequestInformations(RequestInterface $request)
+    private function collectRequestInformations(RequestInterface $request, Stack $stack = null)
     {
-        if (!$stack = $this->collector->getCurrentStack()) {
+        if (!$stack) {
             return;
         }
 
-        $stack = $this->collector->getCurrentStack();
         $stack->setRequestTarget($request->getRequestTarget());
         $stack->setRequestMethod($request->getMethod());
+        $stack->setRequestScheme($request->getUri()->getScheme());
+        $stack->setRequestHost($request->getUri()->getHost());
+        $stack->setClientRequest($this->formatter->formatRequest($request));
+    }
+
+    /**
+     * @param ResponseInterface $response
+     * @param Stack|null        $stack
+     */
+    private function collectResponseInformations(ResponseInterface $response, Stack $stack = null)
+    {
+        if (!$stack) {
+            return;
+        }
+
+        $stack->setResponseCode($response->getStatusCode());
+        $stack->setClientResponse($this->formatter->formatResponse($response));
+    }
+
+    /**
+     * @param \Exception $exception
+     * @param Stack|null $stack
+     */
+    private function collectExceptionInformations(\Exception $exception, Stack $stack = null)
+    {
+        if ($exception instanceof HttpException) {
+            $this->collectResponseInformations($exception->getResponse(), $stack);
+        }
+
+        if (!$stack) {
+            return;
+        }
+
+        $stack->setClientException($this->formatter->formatException($exception));
     }
 }