+++ Collector/ProfileClientFactory.php
@@ -27,16 +27,23 @@ class ProfileClientFactory implements ClientFactory
     private $collector;
 
     /**
+     * @var Formatter
+     */
+    private $formatter;
+
+    /**
      * @param ClientFactory|callable $factory
      * @param Collector              $collector
+     * @param Formatter              $formatter
      */
-    public function __construct($factory, Collector $collector)
+    public function __construct($factory, Collector $collector, Formatter $formatter)
     {
         if (!$factory instanceof ClientFactory && !is_callable($factory)) {
             throw new \RuntimeException(sprintf('First argument to ProfileClientFactory::__construct must be a "%s" or a callable.', ClientFactory::class));
         }
         $this->factory = $factory;
         $this->collector = $collector;
+        $this->formatter = $formatter;
     }
 
     /**
@@ -50,6 +57,6 @@ class ProfileClientFactory implements ClientFactory
             $client = new FlexibleHttpClient($client);
         }
 
-        return new ProfileClient($client, $this->collector);
+        return new ProfileClient($client, $this->collector, $this->formatter);
     }
 }